pyproject.toml
[tool.poetry]
name = "weaviate-agent-demo"
version = "0.1.0"
description = "Demo project for sometimes (poorly) generating Python code for Weaviate"
authors = ["JP Hwang <me@jphwang.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "weaviate_agent_demo", from = "src"}]  # Add this line

[tool.poetry.dependencies]
python = "^3.9"
weaviate-client = "^4.7.1"
claudette = "^0.0.9"
click = "^8.1.7"
streamlit = "^1.38.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
